Fetch per-user information too from GitHub GraphQL API
We need more detailed user information such as email, biography,
follower/following counts, etc.

I've used the GraphQL API instead of the REST one we were using before
because it means we can select the exact information we need.

In future we could probably just do the entire query with GraphQL but
this works for now!